Extend wordcodec benchmarks
